[refactor] refactor domain repo position
