+++ backend/src/domain/entities/quiz.py
@@ -1,4 +1,5 @@
 from typing import List
+import uuid
 from pydantic import Field
 from pydantic.dataclasses import dataclass
 
@@ -7,6 +8,7 @@ from src.domain.entities.question import Question
 
 @dataclass(frozen=True)
 class Quiz:
+    id: str = uuid.uuid4().hex
     questions: List[Question] = Field(
         ..., description="Questionのリスト", min_length=3, max_length=20
     )
